+++ hyperbench/tests/mock/mock.py
@@ -23,7 +23,7 @@ def new_mock_trainer() -> MagicMock:
 
 
 def new_mock_pyg_node2vec() -> MagicMock:
-    weight = torch.nn.Parameter(torch.tensor(1.0))
+    weight = torch.nn.Parameter(torch.tensor(1.0, dtype=torch.float))
 
     model = MagicMock()
     model.to.return_value = model
@@ -34,23 +34,28 @@ def new_mock_pyg_node2vec() -> MagicMock:
         )
     ]
     model.parameters.return_value = [weight]
-    model.return_value = (weight * torch.ones((4, 2))).requires_grad_(True)
+    model.return_value = (weight * torch.ones((4, 2), dtype=torch.float)).requires_grad_(True)
 
     def loss(positive_random_walk: Tensor, negative_random_walk: Tensor) -> Tensor:
-        return weight * (positive_random_walk.float().sum() + negative_random_walk.float().sum())
+        return weight * (
+            positive_random_walk.float().sum(dtype=torch.float)
+            + negative_random_walk.float().sum(dtype=torch.float)
+        )
 
     model.loss.side_effect = loss
     return model
 
 
 def new_mock_villain() -> MagicMock:
-    weight = torch.nn.Parameter(torch.tensor(1.0))
+    weight = torch.nn.Parameter(torch.tensor(1.0, dtype=torch.float))
 
     model = MagicMock()
     model.to.return_value = model
     model.parameters.return_value = [weight]
-    model.hyperedge_embeddings.return_value = torch.ones((2, 3), requires_grad=True)
-    model.node_embeddings.return_value = torch.ones((4, 2), requires_grad=True)
+    model.hyperedge_embeddings.return_value = torch.ones(
+        (2, 3), requires_grad=True, dtype=torch.float
+    )
+    model.node_embeddings.return_value = torch.ones((4, 2), requires_grad=True, dtype=torch.float)
 
     def loss(hyperedge_index: Tensor, num_hyperedges: int) -> tuple[Tensor, dict[str, Any]]:
         return weight * 2.0, {}
